fix: split util
