+++ apps/core/src/common/interceptors/analyze.interceptor.ts
@@ -26,9 +26,9 @@ import { OptionModel } from '~/modules/configs/configs.model'
 import { CacheService } from '~/processors/redis/cache.service'
 import { getNestExecutionContextRequest } from '~/transformers/get-req.transformer'
 import { InjectModel } from '~/transformers/model.transformer'
-import { scheduleManager } from '~/utils'
 import { getIp } from '~/utils/ip.util'
 import { getRedisKey } from '~/utils/redis.util'
+import { scheduleManager } from '~/utils/schedule.util'
 
 @Injectable()
 export class AnalyzeInterceptor implements NestInterceptor {
