+++ apps/core/src/common/interceptors/idempotence.interceptor.ts
@@ -20,7 +20,9 @@ import {
 } from '~/constants/meta.constant'
 import { REFLECTOR } from '~/constants/system.constant'
 import { CacheService } from '~/processors/redis/cache.service'
-import { getIp, getRedisKey, hashString } from '~/utils'
+import { getIp } from '~/utils/ip.util'
+import { getRedisKey } from '~/utils/redis.util'
+import { hashString } from '~/utils/tool.util'
 
 const IdempotenceHeaderKey = 'x-idempotence'
