+++ apps/core/src/modules/comment/comment.service.ts
@@ -29,7 +29,8 @@ import { BarkPushService } from '~/processors/helper/helper.bark.service'
 import { EmailService } from '~/processors/helper/helper.email.service'
 import { EventManagerService } from '~/processors/helper/helper.event.service'
 import { InjectModel } from '~/transformers/model.transformer'
-import { getAvatar, hasChinese, scheduleManager } from '~/utils'
+import { scheduleManager } from '~/utils/schedule.util'
+import { getAvatar, hasChinese } from '~/utils/tool.util'
 
 import { ConfigsService } from '../configs/configs.service'
 import { createMockedContextResponse } from '../serverless/mock-response.util'
