+++ apps/core/src/modules/dependency/dependency.controller.ts
@@ -7,7 +7,7 @@ import { ApiController } from '~/common/decorators/api-controller.decorator'
 import { Auth } from '~/common/decorators/auth.decorator'
 import { HTTPDecorators } from '~/common/decorators/http.decorator'
 import { DATA_DIR } from '~/constants/path.constant'
-import { installPKG } from '~/utils'
+import { installPKG } from '~/utils/system.util'
 
 import { ServerlessService } from '../serverless/serverless.service'
