+++ apps/core/src/modules/health/sub-controller/log.controller.ts
@@ -15,8 +15,8 @@ import { Auth } from '~/common/decorators/auth.decorator'
 import { HTTPDecorators } from '~/common/decorators/http.decorator'
 import { LOG_DIR } from '~/constants/path.constant'
 import { AdapterResponse } from '~/types/request'
-import { formatByteSize } from '~/utils'
 import { getTodayLogFilePath } from '~/utils/path.util'
+import { formatByteSize } from '~/utils/system.util'
 
 import { LogQueryDto, LogTypeDto } from '../health.dto'
