+++ apps/core/src/modules/serverless/serverless.service.ts
@@ -36,14 +36,12 @@ import { EventManagerService } from '~/processors/helper/helper.event.service'
 import { HttpService } from '~/processors/helper/helper.http.service'
 import { CacheService } from '~/processors/redis/cache.service'
 import { InjectModel } from '~/transformers/model.transformer'
-import {
-  getRedisKey,
-  safePathJoin,
-  safeProcessEnv,
-  scheduleManager,
-} from '~/utils'
 import { EncryptUtil } from '~/utils/encrypt.util'
+import { getRedisKey } from '~/utils/redis.util'
 import { safeEval } from '~/utils/safe-eval.util'
+import { scheduleManager } from '~/utils/schedule.util'
+import { safeProcessEnv } from '~/utils/system.util'
+import { safePathJoin } from '~/utils/tool.util'
 
 import { ConfigsService } from '../configs/configs.service'
 import { SnippetModel, SnippetType } from '../snippet/snippet.model'
