+++ apps/core/src/processors/gateway/gateway.service.ts
@@ -8,7 +8,8 @@ import type {
 import { Injectable } from '@nestjs/common'
 
 import { RedisKeys } from '~/constants/cache.constant'
-import { getRedisKey, safeJSONParse } from '~/utils'
+import { getRedisKey } from '~/utils/redis.util'
+import { safeJSONParse } from '~/utils/tool.util'
 
 import { CacheService } from '../redis/cache.service'
