+++ apps/core/src/processors/helper/helper.tq.service.ts
@@ -3,7 +3,7 @@ import type { Redis } from 'ioredis'
 
 import { Injectable } from '@nestjs/common'
 
-import { safeJSONParse } from '~/utils'
+import { safeJSONParse } from '~/utils/tool.util'
 
 import { CacheService } from '../redis/cache.service'
