--- apps/core/src/utils/index.ts
@@ -1,9 +1,0 @@
-export * from './ip.util'
-export * from './jsonschema.util'
-export * from './pic.util'
-export * from './redis.util'
-export * from './system.util'
-export * from './time.util'
-export * from './tool.util'
-export * from './biz.util'
-export * from './schedule.util'
