+++ apps/core/test/src/utils/pic.util.spec.ts
@@ -1,5 +1,4 @@
-import { sleep } from '~/utils'
-import { getAverageRGB, pickImagesFromMarkdown } from '~/utils/pic.util'
+import { pickImagesFromMarkdown } from '~/utils/pic.util'
 
 describe('src/utils/pic.util', () => {
   test('marked ast', async () => {
@@ -17,11 +16,11 @@ describe('src/utils/pic.util', () => {
     ])
   })
 
-  test('get color from image buffer', async () => {
-    const base64 = `iVBORw0KGgoAAAANSUhEUgAAADQAAAA0CAYAAADFeBvrAAAAAXNSR0IArs4c6QAAAERlWElmTU0AKgAAAAgAAYdpAAQAAAABAAAAGgAAAAAAA6ABAAMAAAABAAEAAKACAAQAAAABAAAANKADAAQAAAABAAAANAAAAABdv+0DAAAQj0lEQVRoBb2aaYzV13mH33tngBlmgGEdgzFm383i4BJvYBs7lkiTD20Sb6kSpfnQOFKqSq2SOFXS2G3zgX6oVLuVUndLaKoqVWOrLrGd0NgQvIDNYjD7OuwwDGC22bi3z/M/c+F6mNWt80pn/ts573nX33nPuZMr/sGz0SvlcxGFYq/d+twhB7+IZVEsfoXrXbQ9tCdpe2nXyX7F/s1beX30b+xuMEI+HxX5x+KDKxEXaKOGToyBA7ZGoTAPKVTuI1P+I4/8aANnRy73fhTjsTh4KmLEkIhPLQzeRbS1VcHypTK21WX3fb79TSo0O/L5d6K5dWI0oMyy+RHffyLia5+O+OziiJPnFXo67XnaG7QjePJdrstofaa+hNwiGH8OjrfStNoHtAO01bQ1tL7Q2Mjn1sflluo4dS7iqw9H/I6p00HNhYhKbZvlzO/HxUvkDo9Da0bgvV8y/1CeLnT07vHSm0LzYbgBhhHnmOTqVSZmSM2giAGV3+X9Jrh/n/Zij7Pkcq9FS3tNHG+K+MZnIpbfkbq3tkdsOxmx81hE1YDIFLmKcrMn8zwwYtehiIoK9XyUAX/f4xwdH7sPOYwFfTZDtwKTTBsXcftUgoJrJZOcIURaWhdi+Rfo9680rdgV/QUvp8fRxoivPHRdmXZ4rtuHv5sRGEM14rlRdXy/O+K374lYejuGIyCaW+SJS/tG3XtIl5OqceFyxNgRKd5N3nOg0ms7Iw6fwLqgrIrVDXmc/Lgfjz3CmLXZyPRnPAo/FUfORNw5K+IL917/tP5gxHl4jcYOeez6wKKIexYQ1GDDOaLL65jhEafwavWg2dcH9nzXlULjGUJcxFgaGQsZ32ntiDiKJQdjuSULCY1JEb/eErF9f0Rt9VjCZA0efZwR/5aNi3g6WtpSOH1teccrLrsJs+PwGV4TcRkP/daciGG16b4JAzmXSlYT2olwXRAWgSt7ps4KzYDZZixdFZeYyHi+hMvPXoSLLmOii7xvaY04fRYlBhOUWL0eS67dnPrXVP0EpRrovA5ej8cpBFyO9UcPS5LIbxcoV4sHzE0FN08+EAh4LhmO19m3NErNSKrApT1TZ4Uez5RRkZFDIgaRqJmF4XUCwQYjRBvfSh67SDg2w+KueYZdxM/fQGHmrK1ehVJ/E+1XB2X59iAQXaI9KNOOoWvgWaoCBJvOpP1K35NnQJDeqbNC4+M8lppYH/GnjyakOcvzu4ciNh5OlqzAS4MYljmM+3bmMaFFJq37EinU0jY0BlZ+h5W/EHW1+ThP2EmimiFbrkz6cuNf+bcyzitmonUwyZ67/aNjy6kis95gPGyzhmvEC61wHYi3fFboNEkaV3rOlJpEPsxlxcAIhlLj+XwWurtPAx68a6SVPFw+a+d7UfQKoSlPjde5xuvcv+y5s4dOZIKbM6V4bgRxVBs9rsV3GYNrt0L7OcbdAZo1HGdtaYhYMCVi0TQSnvzZcCgJV2uoXRt1443zmpt7GH8MqK8FgHpb58q4dFZoX1Qz4XES3hXdZFfQvpCeagYsaoDhWSgi9D7ze2mkUL92D/d62nDtQSMBwn47DqS59XSxuLIvIthH25fTKzEAhq49r21N752gJwHKR6vUWRS5FcS/uwwIqlFCRSqYTkAwl7oitygjMMhOvLnncMQQYL1YfI6uIFLfqLNCDRhnVYwhRP4TxGqibJuGcN0JUJpDRRSmmby9jJcGYoS6uoiV/wMPFNzBuiNySubHENDSZ4ZdIyNBpDRv1my0tNIAMItvXuvTh5vOCjnkyWxBU4m/+28EYNY6Fr2eNnglYW+/BVCYmNBpFEZpYs1ym9CCsEL9FRSeNoaKYCpljgspzwKNcxW4niLUV60D04iQ2sHNeOde5AFJ+k5dKXQI630vxlHuvLM34l9eoQRBuPLY1yPWYi0IYji6Vg3D6uPwSj0hUwtC6pm5UyImoAAAGVfp53ojZEvNjFUZvVoHCMweR8XBfLsJtTp4FAq76IVmgfYB875RVwo58mlC7/UYjYC/3hbx6tsI7ELbIYwWNdfqQCAVMzwutIBuTWlWgaWNPn7LMYXeUWnrNg2zZg+5huENTfsswLO3YLQv35fqN3O4UJiLsUxkOgdaxi9pf05bQuuWulNIyz9MvB+P4SjyJnxXr8e6KGAdpzKG1l1TIiaNSlY2JNcfZBFuiJhRT57QzzJJ5a04LqLwdLxl2J3Ge/I4i+ATiIRqXQjJW4QTONraK7K16DxLwbkLo+LSlWW8/w4dXqftoH2dBpMPE+bqllpQajHW30iyjooN2wkjwGb+zLS2GFaSwh88k4T33SHubwHu77iV1WMLgiGcISoQDKVtOoz1GWcI3j01vdt6FGMx1qp86SIUxMNFOjnWfHJ9UzHrR1F0YOVMjPssfP8YGb8Ft39XFCnXh1Ofegb+hIEPZGuLG7y5CPLF+yJuJiRPMMEgHH0eb6jMaSY2lO6bwXqGAayozR03bBamv8Aw826GB8290GYUvMBVSHeTN7wDgLBBVm1oDMk17iyoe5yqY+8RgvBkSoEajFQo/i099FifFLLfWNpqknhWFkJayTXCsn/6RHIDIUbAeJhhhtBWBdPro3BTbbS8/F5m6YrZN8dAQ1YgmDI64v1jCIZwg1FUMsxcRFWqq3Uv+0bfKjzpps+1ah0RcPkKRhiiUj+Cy5cq/mzR8oxfN3/wf6zAQytAvvq40lqZWcqQacT6G3cliw1CmYvkShPhYWjNGRcFCsvLK9+IolX6ldZo34QAeLRyOrZ5BS8dY70xRA09DTKL9+ZVE2Ch8Oj9IVJJgcb9kzQJVJxMO4ZRzuC56kHzSb+dxE+39CyMv54pcBqhLFaFb7cV40fibp4rseZlrIXAKJ2Q60BjxL3TotX1h3O3vFCOMBXkRdtb+2JAXU3kb5+QwjKDfIQUGKQx8DZsXWQNwa4oWzLwpmuWnvnCQxE/+xWnRiDs4KofdKXQWIR7GSHmxRG0H8Ia4QZt6W1Ykbh3ATxOyJkTI/GGdA6lLuF6LayQhF5x476OZ77zKqsQULxoogvh5o0IeAljWOuNR/FT5J8hqYd6IudwLpUucHUn4DvgqLNCk2H4NhOOimNofM/siCfupzYbk9i7ldh5HDQiXMaxpiyZlt5byLrIOoGCslAOmDEu2rdQSaFEjry4evRsVM6/JSomw2szSW3IuTUQ3c7B9wgWl6p4l8mWHm/4mxmHPpZQ2zCa6CsCWqEXiyvKUW4kyuwmfEbECZh/6cGIR+5N/GSy/kDanJm0rhtadyhJ7vpzBoYqI2oJzZ+cnFmwdcvBaNt6OIochuQBharbJkT+CuP0huMPYTQ3fC68kousvBU2WTy9L//rIi5qvvwmoHI4KVJTfZX+P6Pb5/naQbncr5hkROaZJz+dTjP95MTr9qf8EMWIiEwJGauIMa2SopuJLSQfRtBtx2PgQ7OicsJovNQSFdZ2jv8vkMk8m01f17CJ5ON+ng1Fv58kwb0vKdkh3rWLdaNIZyFraHosUCy+z/fP26fDNPHX3N8Wh2H85WXXlZHx2r24lBi33pL0Vqmg1JKSlrVqmIOQmxrYeuwh0YdnSua3n4yKSQi+9UgqjT5FGFvHbeH5TQzViFFU7G7GzwW15o1PEN6dh/SORn5occTMSYQr+ZzPTUeKOxVFhaaj6R/GUdBlyZyIR5f6PtFbhJk5YRiVT2C+uH58EoaG12KuKrl6J94hXEfWYPlRgAfo6KGJvzAMxyAbD6VwWtghtGvPfrx5gLArUWld0nBdkRHhWYMQ/jBKjWNNu9QsOj1ndxX6dgbN1l4enJdoO8l/CvebJ+XK+N09j1Cr1xS+pAxzZd6y9NGVB/G4AppnopLfN+JBK4wZNyW020RSb9gB2JyEb4eCFrDy7IqUxciwUDb0PLvTaxELaW4lib2TWOixpans8BMRFKdBHrcQwrGSlPO32DTWFdw49mjKbYGnnXpmKlbbhYCtTGQZpEIKKNTrtUv1Kdcm02/1Bn5r2JIWaa3tYjmB7wqqNxTW8ZLA4w5aDwkIb20lj+Hn0lKMn9OjRYVqYhgvdhHTHqYr7Os7GIBA6jtzYkrQ0iaOARkSuaJ7Nq2goptJb521ZIo9kjAaQWGysTyUPOUeSNLKC2cCOBj0EgbcRu7tPUTlgKHqMabbcY2kTI71gNMDySMYsOFE4jGUCKH2oIhbblCo0J/ww9OK2E+IfXdlsoD7kSYmsVZbPJc84LmcdLtriMWm92PxyoGjHAC/Smhh0SXzEpAIwRcRwvJG1MqQDAE9CnYL8QHjJwMk2xl/mvlGDiuifC4aycOjCK33DS8RVYWyPRa54ztPg/L5dub/S4z6vSyqkFGF/orCcCS/xXwryyUT1Xj3oMPDc5nIrDP5yn7+YqDFPGCUfroOJRF6mJbDK6OB13ohlnv3P0K9/AQPw1FPWJfJo7aqnX7P4ZWJtHtgMCozmP3pmm0ODUMwkvYi3/6Ba4MvSqRC0rf5uArXfoO4/VycwVqfmJVc76FFYpJ6+levGMsjCbODePbF1xKUGiJZX76b2E0I76YuHRbiDTxhReB6U0IzK2eNt4Vwa7s6AOtvZIY/otEp5sBvXFTksEh2UA8UZz8sH+DaJZUU8qMmXp6V5jex2PnLggf25cqoiGHgLwVafNMu1px30rqhR1wfJItLx7lRsxK4CcUNuQPIYz6Ukyeketk5DdshNd/k849pV2gw7x+VK+TI3822vVMngHBY00paS0sKqBIKbUJq0d2HEnINRcFCgcTMPQgqVWTIZF/rMiuINzrAQzAwrzRMiQQMi0uBwLM4YIhGHEaTD/2lzgq1Zxs39xgvrknluZOpjBMbHiewsgoJnZmn8mdR5kEm3oPCH2RHUy6mhqSxb554nq0SGqdcGaWVtx4dSbim7xW8W8iX1X7uL3VWaAWJ/o+ZsO/hAUnslzSqa4IIY/3k7tOkLBaf4gsZTczncs30qcrKEQWVHNddXZZ14I/GKofnilx96VN/r50V+iesSvGV/yrW/wTMJmQCyTXJBwbH+7RfoMg/c91BK5Exv5OxC7K9vx5Uqc4eKfUuv+pJc9PmPZlW/rk/950Vcuyqjub9VBQhW4Ps5gQznz+GgCd6EHIt5ciC7HTGkxoXPcO0N9JYKp4pk1muA116G3jj9454uvFDx5u9XN+mraUJpyRPj/RStgiqzElyzW17XzxkGHtWYCXgD2q9z9OtEL0p1O3Abj68yvuzWajtO5IW5VIOOsDkN//S7jKxUGEV8sxNBM35G3+sTx/7//f/WyEl+GEm8IFjrCssum7EDCUF9+hLRXcdTDCtIgNAUUurhpNpCUjKnO+/KmnEx6HQigwZVeCd7clLhp6eMpreBUdeeD1VyiLbeCqJnShoYZt+xn/+oyrjuN7O5T7MuwTFH37b+Um0KwIO92d55GI6a1KqnLeSki7IFpa7GyhgL1APABoqqRcHVDYy9onODHt/zvIu69YVyvU+vvceT6PTI0D/7Kw8stxZOANlUEho1hNuHbLwO5TWNHOrUPgirNGsn1QGPB+XQor1ANXCe1TdY2Lz7vTfJnph6GC98QNEPsfi/Ay5NTADkULhGd690k9Vbuj+8SkEcCPsQnLnP9ip3pmdXad9zQtI8VSHJCtR5jPcW7f9tOPd/+nyv/sMPXgp2xIZAAAAAElFTkSuQmCC`
-    const buffer = Buffer.from(base64, 'base64')
-    const hex = await getAverageRGB(buffer, 'image/png')
+  // test('get color from image buffer', async () => {
+  //   const base64 = `iVBORw0KGgoAAAANSUhEUgAAADQAAAA0CAYAAADFeBvrAAAAAXNSR0IArs4c6QAAAERlWElmTU0AKgAAAAgAAYdpAAQAAAABAAAAGgAAAAAAA6ABAAMAAAABAAEAAKACAAQAAAABAAAANKADAAQAAAABAAAANAAAAABdv+0DAAAQj0lEQVRoBb2aaYzV13mH33tngBlmgGEdgzFm383i4BJvYBs7lkiTD20Sb6kSpfnQOFKqSq2SOFXS2G3zgX6oVLuVUndLaKoqVWOrLrGd0NgQvIDNYjD7OuwwDGC22bi3z/M/c+F6mNWt80pn/ts573nX33nPuZMr/sGz0SvlcxGFYq/d+twhB7+IZVEsfoXrXbQ9tCdpe2nXyX7F/s1beX30b+xuMEI+HxX5x+KDKxEXaKOGToyBA7ZGoTAPKVTuI1P+I4/8aANnRy73fhTjsTh4KmLEkIhPLQzeRbS1VcHypTK21WX3fb79TSo0O/L5d6K5dWI0oMyy+RHffyLia5+O+OziiJPnFXo67XnaG7QjePJdrstofaa+hNwiGH8OjrfStNoHtAO01bQ1tL7Q2Mjn1sflluo4dS7iqw9H/I6p00HNhYhKbZvlzO/HxUvkDo9Da0bgvV8y/1CeLnT07vHSm0LzYbgBhhHnmOTqVSZmSM2giAGV3+X9Jrh/n/Zij7Pkcq9FS3tNHG+K+MZnIpbfkbq3tkdsOxmx81hE1YDIFLmKcrMn8zwwYtehiIoK9XyUAX/f4xwdH7sPOYwFfTZDtwKTTBsXcftUgoJrJZOcIURaWhdi+Rfo9680rdgV/QUvp8fRxoivPHRdmXZ4rtuHv5sRGEM14rlRdXy/O+K374lYejuGIyCaW+SJS/tG3XtIl5OqceFyxNgRKd5N3nOg0ms7Iw6fwLqgrIrVDXmc/Lgfjz3CmLXZyPRnPAo/FUfORNw5K+IL917/tP5gxHl4jcYOeez6wKKIexYQ1GDDOaLL65jhEafwavWg2dcH9nzXlULjGUJcxFgaGQsZ32ntiDiKJQdjuSULCY1JEb/eErF9f0Rt9VjCZA0efZwR/5aNi3g6WtpSOH1teccrLrsJs+PwGV4TcRkP/daciGG16b4JAzmXSlYT2olwXRAWgSt7ps4KzYDZZixdFZeYyHi+hMvPXoSLLmOii7xvaY04fRYlBhOUWL0eS67dnPrXVP0EpRrovA5ej8cpBFyO9UcPS5LIbxcoV4sHzE0FN08+EAh4LhmO19m3NErNSKrApT1TZ4Uez5RRkZFDIgaRqJmF4XUCwQYjRBvfSh67SDg2w+KueYZdxM/fQGHmrK1ehVJ/E+1XB2X59iAQXaI9KNOOoWvgWaoCBJvOpP1K35NnQJDeqbNC4+M8lppYH/GnjyakOcvzu4ciNh5OlqzAS4MYljmM+3bmMaFFJq37EinU0jY0BlZ+h5W/EHW1+ThP2EmimiFbrkz6cuNf+bcyzitmonUwyZ67/aNjy6kis95gPGyzhmvEC61wHYi3fFboNEkaV3rOlJpEPsxlxcAIhlLj+XwWurtPAx68a6SVPFw+a+d7UfQKoSlPjde5xuvcv+y5s4dOZIKbM6V4bgRxVBs9rsV3GYNrt0L7OcbdAZo1HGdtaYhYMCVi0TQSnvzZcCgJV2uoXRt1443zmpt7GH8MqK8FgHpb58q4dFZoX1Qz4XES3hXdZFfQvpCeagYsaoDhWSgi9D7ze2mkUL92D/d62nDtQSMBwn47DqS59XSxuLIvIthH25fTKzEAhq49r21N752gJwHKR6vUWRS5FcS/uwwIqlFCRSqYTkAwl7oitygjMMhOvLnncMQQYL1YfI6uIFLfqLNCDRhnVYwhRP4TxGqibJuGcN0JUJpDRRSmmby9jJcGYoS6uoiV/wMPFNzBuiNySubHENDSZ4ZdIyNBpDRv1my0tNIAMItvXuvTh5vOCjnkyWxBU4m/+28EYNY6Fr2eNnglYW+/BVCYmNBpFEZpYs1ym9CCsEL9FRSeNoaKYCpljgspzwKNcxW4niLUV60D04iQ2sHNeOde5AFJ+k5dKXQI630vxlHuvLM34l9eoQRBuPLY1yPWYi0IYji6Vg3D6uPwSj0hUwtC6pm5UyImoAAAGVfp53ojZEvNjFUZvVoHCMweR8XBfLsJtTp4FAq76IVmgfYB875RVwo58mlC7/UYjYC/3hbx6tsI7ELbIYwWNdfqQCAVMzwutIBuTWlWgaWNPn7LMYXeUWnrNg2zZg+5huENTfsswLO3YLQv35fqN3O4UJiLsUxkOgdaxi9pf05bQuuWulNIyz9MvB+P4SjyJnxXr8e6KGAdpzKG1l1TIiaNSlY2JNcfZBFuiJhRT57QzzJJ5a04LqLwdLxl2J3Ge/I4i+ATiIRqXQjJW4QTONraK7K16DxLwbkLo+LSlWW8/w4dXqftoH2dBpMPE+bqllpQajHW30iyjooN2wkjwGb+zLS2GFaSwh88k4T33SHubwHu77iV1WMLgiGcISoQDKVtOoz1GWcI3j01vdt6FGMx1qp86SIUxMNFOjnWfHJ9UzHrR1F0YOVMjPssfP8YGb8Ft39XFCnXh1Ofegb+hIEPZGuLG7y5CPLF+yJuJiRPMMEgHH0eb6jMaSY2lO6bwXqGAayozR03bBamv8Aw826GB8290GYUvMBVSHeTN7wDgLBBVm1oDMk17iyoe5yqY+8RgvBkSoEajFQo/i099FifFLLfWNpqknhWFkJayTXCsn/6RHIDIUbAeJhhhtBWBdPro3BTbbS8/F5m6YrZN8dAQ1YgmDI64v1jCIZwg1FUMsxcRFWqq3Uv+0bfKjzpps+1ah0RcPkKRhiiUj+Cy5cq/mzR8oxfN3/wf6zAQytAvvq40lqZWcqQacT6G3cliw1CmYvkShPhYWjNGRcFCsvLK9+IolX6ldZo34QAeLRyOrZ5BS8dY70xRA09DTKL9+ZVE2Ch8Oj9IVJJgcb9kzQJVJxMO4ZRzuC56kHzSb+dxE+39CyMv54pcBqhLFaFb7cV40fibp4rseZlrIXAKJ2Q60BjxL3TotX1h3O3vFCOMBXkRdtb+2JAXU3kb5+QwjKDfIQUGKQx8DZsXWQNwa4oWzLwpmuWnvnCQxE/+xWnRiDs4KofdKXQWIR7GSHmxRG0H8Ia4QZt6W1Ykbh3ATxOyJkTI/GGdA6lLuF6LayQhF5x476OZ77zKqsQULxoogvh5o0IeAljWOuNR/FT5J8hqYd6IudwLpUucHUn4DvgqLNCk2H4NhOOimNofM/siCfupzYbk9i7ldh5HDQiXMaxpiyZlt5byLrIOoGCslAOmDEu2rdQSaFEjry4evRsVM6/JSomw2szSW3IuTUQ3c7B9wgWl6p4l8mWHm/4mxmHPpZQ2zCa6CsCWqEXiyvKUW4kyuwmfEbECZh/6cGIR+5N/GSy/kDanJm0rhtadyhJ7vpzBoYqI2oJzZ+cnFmwdcvBaNt6OIochuQBharbJkT+CuP0huMPYTQ3fC68kousvBU2WTy9L//rIi5qvvwmoHI4KVJTfZX+P6Pb5/naQbncr5hkROaZJz+dTjP95MTr9qf8EMWIiEwJGauIMa2SopuJLSQfRtBtx2PgQ7OicsJovNQSFdZ2jv8vkMk8m01f17CJ5ON+ng1Fv58kwb0vKdkh3rWLdaNIZyFraHosUCy+z/fP26fDNPHX3N8Wh2H85WXXlZHx2r24lBi33pL0Vqmg1JKSlrVqmIOQmxrYeuwh0YdnSua3n4yKSQi+9UgqjT5FGFvHbeH5TQzViFFU7G7GzwW15o1PEN6dh/SORn5occTMSYQr+ZzPTUeKOxVFhaaj6R/GUdBlyZyIR5f6PtFbhJk5YRiVT2C+uH58EoaG12KuKrl6J94hXEfWYPlRgAfo6KGJvzAMxyAbD6VwWtghtGvPfrx5gLArUWld0nBdkRHhWYMQ/jBKjWNNu9QsOj1ndxX6dgbN1l4enJdoO8l/CvebJ+XK+N09j1Cr1xS+pAxzZd6y9NGVB/G4AppnopLfN+JBK4wZNyW020RSb9gB2JyEb4eCFrDy7IqUxciwUDb0PLvTaxELaW4lib2TWOixpans8BMRFKdBHrcQwrGSlPO32DTWFdw49mjKbYGnnXpmKlbbhYCtTGQZpEIKKNTrtUv1Kdcm02/1Bn5r2JIWaa3tYjmB7wqqNxTW8ZLA4w5aDwkIb20lj+Hn0lKMn9OjRYVqYhgvdhHTHqYr7Os7GIBA6jtzYkrQ0iaOARkSuaJ7Nq2goptJb521ZIo9kjAaQWGysTyUPOUeSNLKC2cCOBj0EgbcRu7tPUTlgKHqMabbcY2kTI71gNMDySMYsOFE4jGUCKH2oIhbblCo0J/ww9OK2E+IfXdlsoD7kSYmsVZbPJc84LmcdLtriMWm92PxyoGjHAC/Smhh0SXzEpAIwRcRwvJG1MqQDAE9CnYL8QHjJwMk2xl/mvlGDiuifC4aycOjCK33DS8RVYWyPRa54ztPg/L5dub/S4z6vSyqkFGF/orCcCS/xXwryyUT1Xj3oMPDc5nIrDP5yn7+YqDFPGCUfroOJRF6mJbDK6OB13ohlnv3P0K9/AQPw1FPWJfJo7aqnX7P4ZWJtHtgMCozmP3pmm0ODUMwkvYi3/6Ba4MvSqRC0rf5uArXfoO4/VycwVqfmJVc76FFYpJ6+levGMsjCbODePbF1xKUGiJZX76b2E0I76YuHRbiDTxhReB6U0IzK2eNt4Vwa7s6AOtvZIY/otEp5sBvXFTksEh2UA8UZz8sH+DaJZUU8qMmXp6V5jex2PnLggf25cqoiGHgLwVafNMu1px30rqhR1wfJItLx7lRsxK4CcUNuQPIYz6Ukyeketk5DdshNd/k849pV2gw7x+VK+TI3822vVMngHBY00paS0sKqBIKbUJq0d2HEnINRcFCgcTMPQgqVWTIZF/rMiuINzrAQzAwrzRMiQQMi0uBwLM4YIhGHEaTD/2lzgq1Zxs39xgvrknluZOpjBMbHiewsgoJnZmn8mdR5kEm3oPCH2RHUy6mhqSxb554nq0SGqdcGaWVtx4dSbim7xW8W8iX1X7uL3VWaAWJ/o+ZsO/hAUnslzSqa4IIY/3k7tOkLBaf4gsZTczncs30qcrKEQWVHNddXZZ14I/GKofnilx96VN/r50V+iesSvGV/yrW/wTMJmQCyTXJBwbH+7RfoMg/c91BK5Exv5OxC7K9vx5Uqc4eKfUuv+pJc9PmPZlW/rk/950Vcuyqjub9VBQhW4Ps5gQznz+GgCd6EHIt5ciC7HTGkxoXPcO0N9JYKp4pk1muA116G3jj9454uvFDx5u9XN+mraUJpyRPj/RStgiqzElyzW17XzxkGHtWYCXgD2q9z9OtEL0p1O3Abj68yvuzWajtO5IW5VIOOsDkN//S7jKxUGEV8sxNBM35G3+sTx/7//f/WyEl+GEm8IFjrCssum7EDCUF9+hLRXcdTDCtIgNAUUurhpNpCUjKnO+/KmnEx6HQigwZVeCd7clLhp6eMpreBUdeeD1VyiLbeCqJnShoYZt+xn/+oyrjuN7O5T7MuwTFH37b+Um0KwIO92d55GI6a1KqnLeSki7IFpa7GyhgL1APABoqqRcHVDYy9onODHt/zvIu69YVyvU+vvceT6PTI0D/7Kw8stxZOANlUEho1hNuHbLwO5TWNHOrUPgirNGsn1QGPB+XQor1ANXCe1TdY2Lz7vTfJnph6GC98QNEPsfi/Ay5NTADkULhGd690k9Vbuj+8SkEcCPsQnLnP9ip3pmdXad9zQtI8VSHJCtR5jPcW7f9tOPd/+nyv/sMPXgp2xIZAAAAAElFTkSuQmCC`
+  //   const buffer = Buffer.from(base64, 'base64')
+  //   const hex = await getAverageRGB(buffer, 'image/png')
 
-    expect(hex).toBe('#fbbdd6')
-  })
+  //   expect(hex).toBe('#fbbdd6')
+  // })
 })
